Modularize base cost page (split into 3 subcomponents)
